requirements.txt
openai-whisper
sounddevice
pynput
numpy
